separated namespace triangles, now it is an independent file (triangle.cpp/hpp)
