+++ CMakeLists.txt
@@ -17,18 +17,18 @@ string(APPEND CMAKE_EXE_LINKER_FLAGS_DEBUG " -fsanitize=address,undefined -fno-o
 
 find_package(SFML COMPONENTS graphics REQUIRED)
 
-add_executable(Boids src/point.cpp src/bird.cpp src/flock.cpp src/statistics.cpp src/graphic.cpp src/main.cpp)
+add_executable(Boids src/point.cpp src/bird.cpp src/flock.cpp src/statistics.cpp src/graphic.cpp src/triangle.cpp src/main.cpp)
 
 target_link_libraries(Boids PRIVATE sfml-graphics)
 
 # if testing enabled...
- if (BUILD_TESTING)
+if (BUILD_TESTING)
 
-   add_executable(Boids.t src/point.cpp src/bird.cpp src/flock.cpp src/statistics.cpp src/graphic.cpp src/test.cpp)
+    add_executable(Boids.t src/point.cpp src/bird.cpp src/flock.cpp src/statistics.cpp src/graphic.cpp src/triangle.cpp src/test.cpp)
 
-   target_link_libraries(Boids.t PRIVATE sfml-graphics )
+    target_link_libraries(Boids.t PRIVATE sfml-graphics)
 
-   # add executable Boids.t to test lists
-   add_test(NAME Boids.t COMMAND Boids.t)
+    # add executable Boids.t to test lists
+    add_test(NAME Boids.t COMMAND Boids.t)
 
- endif()
+endif ()
